Use ROOT5 and ROOT6 based versions of FairSoft for the Travis CI test.
Install the needed debian package from the debian repository, download the fairsoft debian package from our webserver
and install fairsoft.

Allow to set a site name for CDash.

When set use the environment variable SITE for the CDash site name. This is needed to get a meaningful name
for the Travis CI machines. Set the site name for Travis to Travis CI.