Fix Travis CI installation on Mac OSX.
For Mac OSX install gfortran correctly using brew.
Since on the Travis virtual machines gcc is installed without gfortran we have to
uninstall this version first and then reinstall the gcc package with gfortran.

Fix Travis CI tests on Linux and Mac OSX.

Use Root5 and Root6 based versions of FairSoft for the tests.
When one or more tests are failing the return value of ctest is 0 which signals a ctest
run without any problems.
Checking the ouput or the CDash dashboard this is seen easily. With Travis this such a run
is seen as successful. Correct this behaviour by explicitely failing the ctest run when some
tests fail.

Use new patch of fairsoft for Mac OSX.

Due to problem with the boost installation on Mac OSX a new patch version (nov15p2) of FairSoft has to be
used for the Travis tests.