Implement the possibility to modify the ideal geometry.
Add a call to InitParContainers of the modules. This is needed to read parameters to modify the ideal geometry.
Also call the ModifyGeometry function of the modules. In the module specific implementation of the function the
modifications to the ideal geometry can be defined. It is possible to use parameters from a parameter container.
At the end of the simulation the geometry is reverted back to the ideal geometry which is then saved in the
parameter file.